requirements.txt
Django>=4.2.0
celery>=5.3.0
redis>=5.0.0
dnspython>=2.4.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
gunicorn>=21.2.0
whitenoise>=6.5.0
dj-database-url>=2.1.0
djangorestframework>=3.14.0
django-celery-beat>=2.5.0
requests>=2.31.0
ipwhois>=1.2.0

# Testing dependencies
pytest>=7.4.0
pytest-django>=4.5.2
pytest-celery>=0.0.0
pytest-mock>=3.11.1
pytest-cov>=4.1.0
factory-boy>=3.3.0
